+++ jsite-core/src/main/java/com/jsite/modules/sys/web/OfficeController.java
@@ -17,7 +17,6 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpServletResponse;
 import java.util.List;
@@ -26,7 +25,7 @@ import java.util.Map;
 /**
  * 机构Controller
  ** @author liuruijun
- * @version 2013-5-15
+ * @version 2018-5-15
  */
 @Controller
 @RequestMapping(value = "${adminPath}/sys/office")
@@ -48,46 +47,30 @@ public class OfficeController extends BaseController {
 	@RequestMapping(value = "listData")
 	@ResponseBody
 	public List<Office> listData(Office office, Model model) {
-//        /model.addAttribute("list", officeService.findList(office));
-		if (StringUtils.isBlank(office.getParentId())) {
-			office.setParentId("0");
-	    }
+		if (office.getParent() == null || StringUtils.isBlank(office.getParent().getId())) {
+			office.setParent(new Office("0"));
+		}
 		return officeService.findList(office);
-        //return "modules/sys/officeList";
 	}
 
 	@RequiresPermissions("sys:office:view")
 	@RequestMapping(value = {"list", ""})
-	public String list(Office office, Model model) {
-        //model.addAttribute("list", officeService.findList(office));
+	public String list() {
 		return "modules/sys/officeList";
 	}
-	
+
 	@RequiresPermissions("sys:office:view")
 	@RequestMapping(value = "form")
 	public String form(Office office, Model model) {
-		User user = UserUtils.getUser();
-		/*if (office.getParent()==null || office.getParent().getId()==null){
-			office.setParent(user.getOffice());
+		if (office.getParent() == null) {
+			office.setParent(new Office(Office.getRootId()));
 		}
-		office.setParent(officeService.get(office.getParent().getId()));
-		*/if (office.getAreaId()==null){
+
+		if (office.getAreaId() == null) {
+			User user = UserUtils.getUser();
 			office.setAreaId(user.getOffice().getAreaId());
 			office.setAreaName(user.getOffice().getAreaName());
 		}
-		// 自动获取排序号
-	/*	if (StringUtils.isBlank(office.getId())&&office.getParent()!=null){
-			int size = 0;
-			List<Office> list = officeService.findAll();
-			for (int i=0; i<list.size(); i++){
-				Office e = list.get(i);
-				if (e.getParent()!=null && e.getParent().getId()!=null
-						&& e.getParent().getId().equals(office.getParent().getId())){
-					size++;
-				}
-			}
-			office.setCode(office.getParent().getCode() + StringUtils.leftPad(String.valueOf(size > 0 ? size+1 : 1), 3, "0"));
-		}*/
 		model.addAttribute("office", office);
 		return "modules/sys/officeForm";
 	}
@@ -95,20 +78,15 @@ public class OfficeController extends BaseController {
 	@RequiresPermissions("sys:office:edit")
 	@RequestMapping(value = "save")
 	@ResponseBody
-	public String save(Office office, Model model, RedirectAttributes redirectAttributes) {
-		
-		/*if (!beanValidator(model, office)){
-			return form(office, model);
-		}*/
+	public String save(Office office) {
 		
 		if(office.getIsNewRecord()) {
-			if(StringUtils.isEmpty(office.getParentId())) {
+			if(office.getIsRoot()) {
 				office.setTreeLeaf("1");
 				office.setTreeLevel(0);
 				office.setParentIds("0");
-				office.setParentId("0");
 			}else {
-				Office f= officeService.get(office.getParentId());
+				Office f= officeService.get(office.getParent().getId());
 				office.setParentIds(f.getParentIds()+f.getId()+",");
 				office.setTreeLeaf("1");
 				office.setTreeLevel(f.getTreeLevel()+1);
@@ -119,42 +97,22 @@ public class OfficeController extends BaseController {
 			}
 		}
 		officeService.save(office);
-		/*
-		if(office.getChildDeptList()!=null){
-			Office childOffice = null;
-			for(String id : office.getChildDeptList()){
-				childOffice = new Office();
-				childOffice.setName(DictUtils.getDictLabel(id, "sys_office_common", "未知"));
-				//childOffice.setParent(office);
-				childOffice.setArea(office.getArea());
-				childOffice.setType("2");
-				childOffice.setGrade(String.valueOf(Integer.valueOf(office.getGrade())+1));
-				childOffice.setUseable(Global.YES);
-				officeService.save(childOffice);
-			}
-		}*/
 		return renderResult(Global.TRUE, "保存机构'" + office.getName() + "'成功");
-		//addMessage(redirectAttributes, "保存机构'" + office.getName() + "'成功");
-		//String id = "0".equals(office.getParentId()) ? "" : office.getParentId();
-		//return "redirect:" + adminPath + "/sys/office/list?id="+id+"&parentIds="+office.getParentIds();
 	}
 	
 	@RequiresPermissions("sys:office:edit")
 	@RequestMapping(value = "delete")
 	@ResponseBody
-	public String delete(Office office, RedirectAttributes redirectAttributes) {
-		/*if(Global.isDemoMode()){
-			addMessage(redirectAttributes, "演示模式，不允许操作！");
-			return "redirect:" + adminPath + "/sys/office/list";
-		}*/
-//		if (Office.isRoot(id)){
-//			addMessage(redirectAttributes, "删除机构失败, 不允许删除顶级机构或编号空");
-//		}else{
-			officeService.delete(office);
-			return renderResult(Global.TRUE,  "删除机构成功");
-			//addMessage(redirectAttributes, "删除机构成功");
-//		}
-		//return "redirect:" + adminPath + "/sys/office/list?id="+office.getParentId()+"&parentIds="+office.getParentIds();
+	public String delete(Office office) {
+		if(Global.isDemoMode()){
+            return renderResult(Global.FALSE,  "演示模式，不允许操作！");
+		}
+		if (office.getIsRoot()) {
+            return renderResult(Global.TRUE,  "删除机构失败, 不允许删除顶级机构或编号空");
+        }
+
+        officeService.delete(office);
+        return renderResult(Global.TRUE,  "删除机构成功");
 	}
 
 	/**
@@ -180,7 +138,7 @@ public class OfficeController extends BaseController {
 					&& Global.YES.equals(e.getUseable())){
 				Map<String, Object> map = Maps.newHashMap();
 				map.put("id", e.getId());
-				map.put("pId", e.getParentId());
+				map.put("pId", e.getParent().getId());
 				map.put("pIds", e.getParentIds());
 				map.put("name", e.getName());
 				if (type != null && "3".equals(type)){
